refactor: move internal traits into Concerns namespace
Relocate the three internal traits into `BitApps\WPDatabase\Concerns`
(Relations, QueriesRelationships, HasEvents) to separate cross-cutting concerns
from the public surface. Pure organization — no public class moved, no logic
change. Relations/QueriesRelationships gain `use` imports for the QueryBuilder
(and Model) types they reference; Model and QueryBuilder import the relocated
traits. PSR-4 resolves the new path automatically.
